Put StopItem in a different file and added the check for delete authorization along with a toast.
