+++ app/src/main/java/com/github/se/wanderpals/ui/screens/trip/agenda/Agenda.kt
@@ -40,24 +40,16 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.se.wanderpals.R
 import com.github.se.wanderpals.model.viewmodel.AgendaViewModel
 import com.github.se.wanderpals.navigationActions
 import com.github.se.wanderpals.ui.navigation.Route
-import com.github.se.wanderpals.ui.theme.WanderPalsTheme
 import java.time.LocalDate
 import java.time.YearMonth
 
 private const val DAYS_IN_A_WEEK = 7
 private const val MAX_ROWS_CALENDAR = 6
-
-@Preview(showSystemUi = true)
-@Composable
-fun AgendaPreview() {
-  WanderPalsTheme { Agenda(AgendaViewModel("", null)) }
-}
 
 /**
  * The main entry point Composable for the Agenda screen. It displays a calendar view that allows
