+++ app/src/main/java/com/github/se/wanderpals/ui/screens/trip/stops/StopsList.kt
@@ -41,7 +41,6 @@ import androidx.compose.ui.unit.sp
 import com.github.se.wanderpals.model.viewmodel.StopsListViewModel
 import com.github.se.wanderpals.navigationActions
 import com.github.se.wanderpals.ui.PullToRefreshLazyColumn
-import com.github.se.wanderpals.ui.screens.trip.agenda.StopItem
 import com.github.se.wanderpals.ui.screens.trip.agenda.DisplayDate
 import com.github.se.wanderpals.ui.screens.trip.agenda.StopInfoDialog
 import java.time.LocalDate
